Make command-line overrides (somewhat) clearer
This commit:

(1) Changes the order of command-line loading so that config loading
happens before command line overrides. This way, config loading does not
itself have to be concerned with being pre-overridden.

(2) Adds overrides to the data structures that configuration saving uses
to save configuration blocks of the same type, so that they can easily
be added in the future.

(3) Corrects some (all?) existing problems with command-line overrides
being ignored.