qb: Fix building Qt without pkgconf
When building without a pkg-config implementation or with
PKG_CONF_PATH=none Qt support will be disabled because the
check_val function is not correctly being used.

The $4 argument is the relative include directory that needs to be
checked. Correcting this exposes the new issue that two include
directories need to be checked while check_val only supports one.

So check_val now supports checking 1 or more include directories
and Qt5 support builds fine without pkg-config now.

If any of the include directories does not exist it will disable the
checked feature.