Add verb for menu command: "Create Game-options file"
This corrects the text that is displayed on the menu for the command to generate a Game-options file. It is an action so there should be a verb.


I noticed there is another enum called `MENU_ENUM_LABEL_VALUE_GAME_SPECIFIC_OPTIONS_CREATE` with the full three words which doesn't seem to be used.