Rewrite file hasher to limit amount hashed
== DETAILS
Since the content file could potentially be huge, hashing the
whole thing at runtime may take a really long time. Plus, it
was loading the whole file into RAM at once.

Now, we only load 1MB at a time and hash up to the first 64MB.

== TESTING
I don't have any large content files to test it with, but I
tested it with a small one and confirmed that the hash was
correct.