Fix crashes with '--disable-menu' and the Qt frontend.
When failing to load content in the companion ui when HAVE_MENU is not
defined RetroArch will crash in just about every input and video driver.
Even if several sanity checks are added the dummy core will immediately
exit.

Now it will print that it failed to load the core in the companion ui
and reinit the dummy core to match the behavior with the null menu
driver.