libretro-common: improve PRI_SIZET for non-windows platforms
In non-windows platfoms, the `size_t` type may not necessarily use
the `%lu` format specification. For example in 32 bits platforms
instead it needs to be `%u`. Therefore, for non-windows platforms,
it is better to define PRI_SIZET more precisely.

Silences these types of warnings in 32 bits non-windows platforms:

    libretro-common/file/config_file.c: In function ‘config_get_size_t’:
    libretro-common/file/config_file.c:694:32: warning: format ‘%lu’ expects
    argument of type ‘long unsigned int *’, but argument 3 has type
    ‘size_t * {aka unsigned int *}’ [-Wformat=]
           if (sscanf(entry->value, "%" PRI_SIZET, &val) == 1)
                                    ^~~

Discussed in #8191