xmb: Prevent crashes when resizing to a tiny window.
If the user resized their RetroArch window while using xmb it would
initiate a context_reset to scale xmb. When the window is too small
this would result in invalid pointers and a crash in the video driver.

This makes it so that when the window is too small it will skip the
context_reset and not crash. The window at this size is not really
useful to scale anyways.

Fixes https://github.com/libretro/RetroArch/issues/8260