intl: Add intl_bom.sh
This script supports testing for missing BOM in c and header
files as well as automatically fixing them.

Travis will now test if these files contain a BOM and fail
when missing. They can be trivially fixed by running this
script.

To print errors when the BOM is missing.

  intl/intl_bom.sh

To automatically fix the files.

  intl/intl_bom.sh -f