Add variant builds
aarch64 is a 64-bit only build that was used for some time so users
on the main builds could use armv7 dynarecs. Keeping for now so
users of this variant won't lose updates.

ra32 is 32-bit only. Google is mandating 64-bit support in all
apps on the play store. So users of the main builds on 64-bit
devices will no longer have access to 32-bit cores. This build will
be available on alternate stores and via sideloading so cores with
only 32-bit dynarecs can be used on 64-bit devices.