Update libretro-db/README.md
Looking through libretro-db/query.c I discovered the b prefix to strings that turns the string into binary, this is useful for searching up hexadecimal hashes.  This information was not described in the documentation, so I've added it before the Names only search to group with the other find commands, though it will work with get-names as well.