Fix for Remote RetroPad input
This commit fixes processing of Remote RetroPad input. None of the remote inputs are being executed both in menu and in game. This is due to the way current key binds are being detected which block processing of any remote input.

It's been tested using Remote RetroPad core on Android including digital dpad and analog control input.