Change URL for Lakka online updater - nightlies
currently, if users try to update a nightly build, they are offered
latest stable release, which in fact is a downgrade.
with correct URL the users will be pointed to the latest nightly build
in case they are running a nightly build - it does not affect users
running a stable/rc release of Lakka.

originally we wanted to patch RA locally
(https://github.com/libretro/Lakka-LibreELEC/pull/752), but it seems
patching upstream and just adding a make flag is more sane.