Spanish language revision and update
Hello there, this is a large update and proofread for RetroArch's Spanish translation.

I have done the following changes:
 - Translated the new strings.
 - Unified the tenses being used for menu entry labels and sublabels (There was a mix of tenses all over the place, so I've decided for now to use infinitive for labels and present's second person in singular for starting sublabels).
 - Unified many terms that were translated in more than one way (frames, for example, had frames and fotogramas all around).
 - Added many missing trailing periods at sublabels and what I think are OSD notifications.
 - Fixed a metric ton of missing accents.
 - Added rewrites on some translations.
 - Updated translations wherever the English texts had substantial changes.

There's one thing that I might have mixed, and that's the Spanish region that the line's being directed. Back when I started this translation, I followed the European Spanish expressions and verbal tenses, but it seems there's been a deviation since then into Latin American Spanish's expressions and verbal usage (video/indice instead of vídeo/índice, encontró instead of ha encontrado... Not sure if it's actually an attempt to do Neutral Spanish).

A decision has to be made about this (Either try to go with Neutral or split the translation into European Spanish and Latin American. I'd personally go with the latter).