Added non-X11 lightgun support
The current method for getting the lightgun aiming coordinates "udev_lightgun_aiming_state" only work in X11 which is shown if you follow the code through, the required coordinates are not set otherwise. I've added an X11 if statement and left that block of code as it is. For the non-X11 code block I've used the methods from the touchscreen/pointer function which successfully gets absolute coordinates back from the mouse device in non-X11. This update is tested and working on the Raspberry Pi for an absolute mouse based device and it was not working before.