Implement ffmpeg send/receive encoding API
This is implementing the send/receive encoding API of ffmpeg in the
ffmpeg record backend.

Newer encoder do need the client to use this API to use them (for
example h264_amf and hvec_amf). "Old" encoders should work fine
(Tested x264 and webm profiles).

The biggest change is the handling of the encoded packets, since
the API can return multiple packets in it's queue. As a positive
side-effect the code for flushing is much more easier.