fix mouse capture hotkey not working
There's a normal binding for this set in your configs/`grab_mouse_toggle`, but
there's a second place where it's hardcoded to check the F11 key. Because F11
is the default key, both of these codepaths will run which immediately captures
and then uncaptures the mouse cursor, effectively breaking it.

Tested and this fixes the behavior on Linux and Emscripten