Save netplay settings which have a CLI override.
Currently the CLI overrides should not be saved when creating a confi-
guration file. But because of a bug, the user can never save netplay
settings which have an override.

This change prevents the frontend to set the "override flag" for these
options when setting the setting manually via the settings menu.

This enables toe saving of the affected settings again.