Search in PREFIX /share/retroarch for default dirs
Some things like default dirs for:

- CORE_INFO
- AUTOCONFIG
- ASSETS
- DATABASE

Are nice to have avilable on initial launch. We can always download
them with the updater, but some distributors might want to install
them directly.

Unfortunately, this means that we can’t use the online updater, since
the prefix is most likely not writable. You can always change this
value to a writable location though. Another way to fix this would be
to only use the prefixed directories when on doesn’t exist in the
$HOME/.config/retroarch directories.