Put local includes first to prevent system headers from interfering
For example, a popular package 'librhash' (dependency of CMake)
installs a header called 'rhash.h' which collides with a local header
of the same name. A system include path provided by the user
would end up in config.mk, which would be the first -I argument,
and that library's header would be included instead of what we expect.

To avoid this problem, prepend local include paths instead of appending.