Support fallback dirs to cycle user shader presets
When a user shader preset was saved, if the Video Shader directory
wasn't writable, it would fall back to the Menu Config directory, and
failing that, it would fall back to the directory where retroarch.cfg
is stored.

This change allows the shader cycling feature to use those directories
if no shader presets are found on the Video Shader directory.