Improving Wifi API and connman implementation
This sets the ground for more PRs to come. Clean up how scanning and listing
SSID works, and adds attributes to the newtworks.
Frontend will be able to know which networks are "remembered" and thus
require no password to re-connect.

For now scans are performed every 2 minutes or on demand if no networks found
(either for real or in the "cache").