Remove the path_is_valid() call when loading textures
== DETAILS
The `path_is_valid()` call ends up invoking `stat()`, which is unpredictable.
It's practically free when implemented by an operating system (Linux, Windows),
but on embedded systems it can be problematic.

In the case of the Wii U, the stat() call actually ends up taking longer than
the file I/O. And the texture loading code already handles a "file not found"
case, so the stat() call ends up just being a waste of time.

Removing this reduces the texture load time from ~210-250ms down to 60-65ms
on average.